Rename request/response classes to better match Reel
This factors the classes used by requests and responses within the namespace of
the Request and Response classes. The names are changed to better match Reel:

* HTTP::RequestStream  => HTTP::Request::Writer
* HTTP::ResponseParser => HTTP::Response::Parser
* HTTP::ResponseBody   => HTTP::Response::Body
