+++ lib/http.rb
@@ -4,10 +4,10 @@ require 'http/chainable'
 require 'http/client'
 require 'http/options'
 require 'http/request'
-require 'http/request_stream'
+require 'http/request/writer'
 require 'http/response'
-require 'http/response_body'
-require 'http/response_parser'
+require 'http/response/body'
+require 'http/response/parser'
 require 'http/uri_backport' if RUBY_VERSION < '1.9.0'
 
 # HTTP should be easy
